requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-openai>=0.3.13
chromadb>=0.4.22
openai>=1.75.0
tiktoken>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
whisper
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.37.0
pybud
python-multipart>=0.0.6
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
# pydub>=0.25.1
# pyannote.audio>=3.1.1
openai-whisper>=20231117
sentencepiece>=0.2.0
protobuf>=4.25.3
soundfile>=0.12.1
langgraph
langgraph.graph
langsmith
numpy<1.26.0,>=1.18.5

# Multi-user authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.6.0
email-validator>=2.0.0
requests>=2.28.2
